Refactor Contacts import for better separation e reusability
